Move eanProduct csv generating into own file
